Updating this stuff

+++ p_char.c
@@ -6,13 +6,13 @@
 /*   By: otodd <otodd@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/01 13:57:40 by otodd             #+#    #+#             */
-/*   Updated: 2023/12/01 14:05:09 by otodd            ###   ########.fr       */
+/*   Updated: 2023/12/04 14:35:55 by otodd            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	p_char(char c)
+size_t	p_char(char c)
 {
 	write(1, &c, 1);
 	return (1);
